extract admin menu as component
