+++ frontend/src/components/Game/Card/ActionBar.vue
@@ -1,10 +1,10 @@
 <script setup>
-import { ref, inject } from "vue";
+import { ref } from "vue";
 import { downloadRomApi } from "@/services/api.js";
 import useDownloadStore from "@/stores/download.js";
+import AdminMenu from "@/components/AdminMenu/Base.vue";
 
 // Props
-const emitter = inject("emitter");
 const props = defineProps(["rom"]);
 const saveFiles = ref(false);
 const downloadStore = useDownloadStore();
@@ -55,41 +55,7 @@ const downloadUrl = `${window.location.origin}${props.rom.download_path}`;
             variant="text"
           />
         </template>
-        <v-list rounded="0" class="pa-0">
-          <v-list-item
-            @click="emitter.emit('showSearchDialog', rom)"
-            class="pt-4 pb-4 pr-5"
-          >
-            <v-list-item-title class="d-flex"
-              ><v-icon icon="mdi-search-web" class="mr-2" />Search
-              IGDB</v-list-item-title
-            >
-          </v-list-item>
-          <v-divider class="border-opacity-25" />
-          <v-list-item
-            @click="emitter.emit('showEditDialog', rom)"
-            class="pt-4 pb-4 pr-5"
-          >
-            <v-list-item-title class="d-flex"
-              ><v-icon
-                icon="mdi-pencil-box"
-                class="mr-2"
-              />Edit</v-list-item-title
-            >
-          </v-list-item>
-          <v-divider class="border-opacity-25" />
-          <v-list-item
-            @click="emitter.emit('showDeleteDialog', rom)"
-            class="pt-4 pb-4 pr-5 text-red"
-          >
-            <v-list-item-title class="d-flex"
-              ><v-icon
-                icon="mdi-delete"
-                class="mr-2"
-              />Delete</v-list-item-title
-            >
-          </v-list-item>
-        </v-list>
+        <admin-menu :rom="rom" />
       </v-menu>
     </v-row>
   </v-card-text>
